scripts/train_model.py
# scripts/train_model.py
"""ML model training pipeline."""
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error, r2_score
import pickle


def train_crop_yield_model(data_path: str, model_output_path: str):
    """Train and save crop yield prediction model."""
    # TODO: Load data, preprocess, train model
    pass


if __name__ == "__main__":
    train_crop_yield_model("data/crop_data.csv", "ml_models/crop_yield_model.pkl")

