requirements.txt
# requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1
flask-cors==4.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# ML
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.2

# API & HTTP
requests==2.31.0
python-dotenv==1.0.0

# Validation
pydantic==2.5.2
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
